pyproject.toml
[project]
name = "ragnest"
version = "0.1.0"
description = "Multi-knowledge-base RAG system with MCP integration"
requires-python = ">=3.12"
license = "MIT"
readme = "README.md"
authors = [{ name = "november-pain" }]
keywords = ["rag", "mcp", "embeddings", "knowledge-base", "pgvector", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "fastmcp>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "ollama>=0.4",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "langchain-text-splitters>=0.3",
    "pyarrow>=15.0",
    "pandas>=2.0",
    "rich>=13.0",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "basedpyright>=1.20",
    "pre-commit>=4.0",
    "pip-audit>=2.7",
    "bandit>=1.7",
    "types-psycopg2>=2.9",
    "types-PyYAML>=6.0",
    "pandas-stubs>=2.0",
]

[project.scripts]
ragnest-worker = "ragnest.cli.worker:main"
ragnest-db = "ragnest.cli.db_setup:main"
ragnest-server = "ragnest.mcp.server:main"

[project.urls]
Homepage = "https://github.com/november-pain/ragnest"
Repository = "https://github.com/november-pain/ragnest"
Issues = "https://github.com/november-pain/ragnest/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ragnest"]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = [
    "E", "F", "W",       # pycodestyle + pyflakes
    "I",                  # isort
    "N",                  # pep8-naming
    "UP",                 # pyupgrade
    "ANN",                # annotations
    "B",                  # bugbear
    "A",                  # builtins shadowing
    "COM",                # commas
    "C4",                 # comprehensions
    "DTZ",                # datetime
    "ISC",                # string concat
    "ICN",                # import conventions
    "PIE",                # misc lints
    "PT",                 # pytest style
    "RSE",                # raise
    "RET",                # return
    "SLF",                # private access
    "SIM",                # simplify
    "TID",                # tidy imports
    "TCH",                # type checking
    "ARG",                # unused args
    "PTH",                # pathlib
    "PL",                 # pylint
    "TRY",                # tryceratops
    "FLY",                # flynt
    "PERF",               # perflint
    "RUF",                # ruff-specific
]
ignore = ["ANN401", "TRY003", "COM812", "ISC001", "PLR0913", "PLR0915"]

[tool.ruff.lint.isort]
known-first-party = ["ragnest"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["fitz", "pgvector.*"]
ignore_missing_imports = true

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
markers = [
    "integration: marks tests that need a database",
]
